# Conflicts:
#	app/src/main/java/com/example/android/employeesmanagementapp/activities/AddTaskActivity.java

+++ app/src/main/java/com/example/android/employeesmanagementapp/data/entries/TaskEntry.java
@@ -121,7 +121,7 @@ public class TaskEntry {
         return taskRating;
     }
 
-    public void setTaskRating(int taskRating) {
+    public void setTaskRating(float taskRating) {
         this.taskRating = taskRating;
     }
